com.unity.textmeshpro@3.2.0-pre.10 ## [3.2.0-pre.10] - 2024-05-27 ### Changes - Ensure space and underline are always added to Static FontAsset. (UUM-45512) - Fix bug that occurs with the Input Field, resulting in an incorrect cursor position when modifying long text input. (UUM-58685) - Fixed compile error on TMP_PostBuildProcessorHandler.cs when building for iOS with "install into source code (UUM-57710) - Fixed opening style tag overflow and TextSettings upgrade (UUM-30205) - Fixed crash occurring in the FontAssetCreatorWindow. (UUM-66950) - Fixed incorrect handling of Emoji Presentation forms ensuring emojis such as ©️ are displayed in text form, whereas 🚀 is displayed in its color presentation form. - Added support for Emoji Variant Selectors u+FE0E and u+FE0F to control whether an emoji is displayed in text or presentation form: GitHub PR #48504

+++ Scripts/Editor/TMP_FontAsset_CreationMenu.cs
@@ -227,7 +227,7 @@ namespace TMPro
             AssetDatabase.AddObjectToAsset(mat, fontAsset);
 
             // Add Font Asset Creation Settings
-            fontAsset.creationSettings = new FontAssetCreationSettings(fontAsset.m_SourceFontFileGUID, fontAsset.faceInfo.pointSize, 0, atlasPadding, 0, 1024, 1024, 7, string.Empty, (int)renderMode);
+            fontAsset.creationSettings = new FontAssetCreationSettings(fontAsset.m_SourceFontFileGUID, (int)fontAsset.faceInfo.pointSize, 0, atlasPadding, 0, 1024, 1024, 7, string.Empty, (int)renderMode);
 
             // Not sure if this is still necessary in newer versions of Unity.
             //EditorUtility.SetDirty(fontAsset);
